Move funnel data types next to calculateFunnelSteps in visualizations
Step, Measure, FunnelDatum, and FunnelStep were defined in static-viz's
FunnelChart types but are the vocabulary of calculateFunnelSteps, which
lives in visualizations/lib/funnel. Move the four types to a new
visualizations/lib/funnel/types.ts and update all importers (funnel
utils, FunnelChart, static-viz funnel/margin utils, and the funnel
spec). FunnelSettings stays in static-viz since it depends on the local
NumberFormatOptions.
