+++ frontend/src/metabase/static-viz/components/FunnelChart/FunnelChart.tsx
@@ -2,10 +2,7 @@ import { Group } from "@visx/group";
 import { Line, Polygon } from "@visx/shape";
 import { Fragment } from "react";
 
-import type {
-  FunnelDatum,
-  FunnelSettings,
-} from "metabase/static-viz/components/FunnelChart/types";
+import type { FunnelSettings } from "metabase/static-viz/components/FunnelChart/types";
 import {
   calculateFunnelPolygonPoints,
   getFormattedStep,
@@ -14,6 +11,7 @@ import {
 } from "metabase/static-viz/components/FunnelChart/utils/funnel";
 import { Text } from "metabase/static-viz/components/Text";
 import { measureTextHeight } from "metabase/static-viz/lib/text";
+import type { FunnelDatum } from "metabase/visualizations/lib/funnel/types";
 import {
   calculateFunnelSteps,
   calculateStepOpacity,
