+++ frontend/src/metabase/static-viz/components/FunnelChart/types.ts
@@ -1,11 +1,6 @@
 import type { NumberFormatOptions } from "metabase/static-viz/lib/numbers";
 import type { VisualizationSettings } from "metabase-types/api";
 
-export type Step = string | number;
-export type Measure = number;
-
-export type FunnelDatum = [Step, Measure];
-
 export type FunnelSettings = {
   step: {
     name: string;
@@ -20,13 +15,4 @@ export type FunnelSettings = {
     border: string;
   };
   visualization_settings: VisualizationSettings;
-};
-
-export type FunnelStep = {
-  step: string | number;
-  measure: number;
-  percent: number;
-  top: number;
-  left: number;
-  height: number;
 };
