+++ frontend/src/metabase/visualizations/lib/funnel/types.ts
@@ -1,0 +1,13 @@
+export type Step = string | number;
+export type Measure = number;
+
+export type FunnelDatum = [Step, Measure];
+
+export type FunnelStep = {
+  step: string | number;
+  measure: number;
+  percent: number;
+  top: number;
+  left: number;
+  height: number;
+};
